moved passenger marker to other file
