+++ client/src/components/map/map.css
@@ -38,7 +38,3 @@
 .map__combobox:hover::placeholder {
     color: var(--color-blue);
 }
-
-.passenger__marker:hover {
-    content: "http://maps.google.com/mapfiles/ms/icons/yellow-dot.png"
-}
